concurrent processing released
