+++ 01_web/http-server/src/main/java/ru/netology/Connection.java
@@ -1,0 +1,86 @@
+package ru.netology;
+
+import java.io.BufferedOutputStream;
+import java.io.BufferedReader;
+import java.io.InputStreamReader;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.time.LocalDateTime;
+import java.util.concurrent.Callable;
+
+public class Connection implements Callable<Boolean> {
+
+    Server server;
+
+    protected Connection(Server server) {
+        this.server = server;
+    }
+
+
+    @Override
+    public Boolean call() throws Exception {
+        try (
+                final var socket = server.serverSocket.accept();
+                final var in = new BufferedReader(new InputStreamReader(socket.getInputStream()));
+                final var out = new BufferedOutputStream(socket.getOutputStream())
+        ) {
+            // read only request line for simplicity
+            // must be in form GET /path HTTP/1.1
+            final var requestLine = in.readLine();
+            final var parts = requestLine.split(" ");
+
+            if (parts.length != 3) {
+                // just close socket
+                return false;
+            }
+
+            final var path = parts[1];
+            if (!server.validPaths.contains(path)) {
+                out.write((
+                        "HTTP/1.1 404 Not Found\r\n" +
+                                "Content-Length: 0\r\n" +
+                                "Connection: close\r\n" +
+                                "\r\n"
+                ).getBytes());
+                out.flush();
+                return false;
+            }
+
+            final var filePath = Path.of(".", "public", path);
+            final var mimeType = Files.probeContentType(filePath);
+
+            // special case for classic
+            if (path.equals("/classic.html")) {
+                final var template = Files.readString(filePath);
+                final var content = template.replace(
+                        "{time}",
+                        LocalDateTime.now().toString()
+                ).getBytes();
+                out.write((
+                        "HTTP/1.1 200 OK\r\n" +
+                                "Content-Type: " + mimeType + "\r\n" +
+                                "Content-Length: " + content.length + "\r\n" +
+                                "Connection: close\r\n" +
+                                "\r\n"
+                ).getBytes());
+                out.write(content);
+                out.flush();
+                return false;
+            }
+
+            final var length = Files.size(filePath);
+            out.write((
+                    "HTTP/1.1 200 OK\r\n" +
+                            "Content-Type: " + mimeType + "\r\n" +
+                            "Content-Length: " + length + "\r\n" +
+                            "Connection: close\r\n" +
+                            "\r\n"
+            ).getBytes());
+            Files.copy(filePath, out);
+            out.flush();
+        }
+
+
+        return false;
+    }
+}
